refactor(cli): extract archive routes into server-routes/archive.ts
#354 pass 3. Moves the archive endpoints (/api/archived, /api/archive/:slug
POST+DELETE) and their directory-marker helpers (getArchivedSlugs/archiveSlug/
unarchiveSlug + ARCHIVE_DIR) out of startServer into a register function,
following the existing registerSessionAssetRoutes pattern. Deps: { baseDir }.

server.ts: 3263 -> 3219. No behavior change — all 340 CLI tests pass,
typecheck + lint clean.

Refs #354
